Density Estimation Tools

Helper functions for working with data for density tensors.

Warning

This submodule is deprecated and may be removed in future versions.

Density Estimation Tools

Matrix Tensor Factor

Block Tensor Decomposition

About this package

Quick Guide

Factorizing your data

Extracting factors

Iteration Statistics

Output keyword arguments

Constants

Functions

Types

Block Update Order

Randomizing Block Updates Order

Constrained Factorization

Example

Ready-to-use Constraints

Advanced Constraints

Manual Constraint Updates

Decomposition Models

How Julia treats an AbstractDecomposition

Iteration Stats

BlockTensorFactorization.DensityEstimationTools.DEFAULT_ALPHA — Constant

DEFAULT_ALPHA = 0.9::Real

Smoothing parameter for calculating a kernel's bandwidth.

BlockTensorFactorization.DensityEstimationTools.DEFAULT_N_SAMPLES — Constant

DEFAULT_N_SAMPLES = 64::Int

Number of samples to use when standardizing a vector of density estimates.

BlockTensorFactorization.DensityEstimationTools._in2drange — Method

Returns a function that checks if each coordinate is in the inner P percentile of the values in vs.

BlockTensorFactorization.DensityEstimationTools._inrange — Method

Returns a function that checks if a value is in the inner P percentile of the values in v.

BlockTensorFactorization.DensityEstimationTools.coordzip — Method

coordzip(rcoords)

Zips the "x" and "y" values together into a list of x coords and y coords. Example –––- coordzip([(0,0), (1,1), (1,1), (1,1), (1,2), (1,2)])

[[0, 1, 1, 1, 1, 1], [0, 1, 1, 1, 2, 2]]

BlockTensorFactorization.DensityEstimationTools.default_bandwidth — Function

default_bandwidth(data; alpha=0.9, inner_percentile=100)

Coppied from KernelDensity since this function is not exported. I want access to it so that the same bandwidth can be used for different densities for the same measurements.

BlockTensorFactorization.DensityEstimationTools.filter_2d_inner_percentile — Method

Filters 2d elements so only the ones in the inner P percentile remain. See filter_inner_percentile.

BlockTensorFactorization.DensityEstimationTools.filter_inner_percentile — Method

Filters elements so only the ones in the inner P percentile remain. See filter_2d_inner_percentile.

BlockTensorFactorization.DensityEstimationTools.kde2d — Method

kde2d((xs, ys), values)

Performs a 2d KDE based on two lists of coordinates, and the value at those coordinates. Input ––-

xs, ys::Vector{Real}: coordinates/locations of samples

values::Vector{Integer}: value of the sample

Returns

f::BivariateKDE use f.x, f.y for the location of the (re)sampled KDE,

and f.density for the sample values of the KDE

BlockTensorFactorization.DensityEstimationTools.make_densities — Method

make_densities(s::Sink; kwargs...)
make_densities(s::Sink, domains::AbstractVector{<:AbstractVector}; kwargs...)

Estimates the densities for each measurement in a Sink.

When given domains, a list where each entry is a domain for a different measurement, resample the kernel on this domain.

Parameters

bandwidths::AbstractVector{<:Real}: list of bandwidths used for each measurement's

density estimation

inner_percentile::Integer=100: value between 0 and 100 that filters out each measurement

by using the inner percentile range. This can help remove outliers and focus in on where the bulk of the data is.

Returns

density_estimates::Vector{UnivariateKDE}

BlockTensorFactorization.DensityEstimationTools.make_densities2d — Method

makedensities2d(s::Sink; kwargs...) makedensities2d(s::Sink, domains::AbstractVector{<:AbstractVector}; kwargs...)

Similar to make_densities but performs the KDE on 2 measurements jointly.

BlockTensorFactorization.DensityEstimationTools.repeatcoord — Method

repeatcoord(coordinates, values)

Repeates coordinates the number of times given by values.

Both lists should be the same length.

Example

coordinates = [(0,0), (1,1), (1,2)] values = [1, 3, 2] repeatcoord(coordinates, values)

[(0,0), (1,1), (1,1), (1,1), (1,2), (1,2)]

BlockTensorFactorization.DensityEstimationTools.standardize_2d_KDEs — Method

standardize_2d_KDEs(KDEs::AbstractVector{BivariateKDE}; n_samples=DEFAULT_N_SAMPLES,)

Resample the densities so they all are sampled from the same x and y coordinates.

BlockTensorFactorization.DensityEstimationTools.standardize_KDEs — Method

Resample the densities within each sink/source so that like-measurements use the same scale.

BlockTensorFactorization.DensityEstimationTools.standardize_KDEs — Method

standardize_KDEs(KDEs::AbstractVector{UnivariateKDE}; n_samples=DEFAULT_N_SAMPLES,)

Resample the densities so they all are sampled from the same domain.

BlockTensorFactorization.MatrixTensorFactor

BlockTensorFactorization.DensityEstimationTools.DEFAULT_ALPHA

BlockTensorFactorization.DensityEstimationTools.DEFAULT_N_SAMPLES

BlockTensorFactorization.MatrixTensorFactor.IMPLIMENTED_CRITERIA

BlockTensorFactorization.MatrixTensorFactor.IMPLIMENTED_METRICS

BlockTensorFactorization.MatrixTensorFactor.IMPLIMENTED_NORMALIZATIONS

BlockTensorFactorization.MatrixTensorFactor.IMPLIMENTED_OPTIONS

BlockTensorFactorization.MatrixTensorFactor.IMPLIMENTED_PROJECTIONS

BlockTensorFactorization.MatrixTensorFactor.IMPLIMENTED_STEPSIZES

BlockTensorFactorization.Core.AbstractConstraint

BlockTensorFactorization.Core.AbstractDecomposition

BlockTensorFactorization.Core.AbstractObjective

BlockTensorFactorization.Core.AbstractStat

BlockTensorFactorization.Core.AbstractStep

BlockTensorFactorization.Core.AbstractTucker

BlockTensorFactorization.Core.BlockGradientDescent

BlockTensorFactorization.Core.CPDecomposition

BlockTensorFactorization.Core.ComposedConstraint

BlockTensorFactorization.Core.ConstraintUpdate

BlockTensorFactorization.Core.DiagonalTuple

BlockTensorFactorization.Core.DisplayDecomposition

BlockTensorFactorization.Core.Entrywise

BlockTensorFactorization.Core.Entrywise

BlockTensorFactorization.Core.EuclidianLipschitz

BlockTensorFactorization.Core.EuclidianStepSize

BlockTensorFactorization.Core.FactorNorms

BlockTensorFactorization.Core.GenericConstraint

BlockTensorFactorization.Core.GenericDecomposition

BlockTensorFactorization.Core.GradientDescent

BlockTensorFactorization.Core.L2

BlockTensorFactorization.Core.L2

BlockTensorFactorization.Core.LinearConstraint

BlockTensorFactorization.Core.LipschitzStep

BlockTensorFactorization.Core.LipschitzStep

BlockTensorFactorization.Core.MomentumUpdate

BlockTensorFactorization.Core.NoConstraint

BlockTensorFactorization.Core.PrintStats

BlockTensorFactorization.Core.ProjectedNormalization

BlockTensorFactorization.Core.Projection

BlockTensorFactorization.Core.Rescale

BlockTensorFactorization.Core.ScaledNormalization

BlockTensorFactorization.Core.SingletonDecomposition

BlockTensorFactorization.Core.SuperDiagonal

BlockTensorFactorization.Core.SuperDiagonal

BlockTensorFactorization.Core.Tucker

BlockTensorFactorization.Core.Tucker

BlockTensorFactorization.Core.Tucker1

BlockTensorFactorization.Core.Tucker1

BlockTensorFactorization.Core.:×₁

BlockTensorFactorization.Core.:×₂

BlockTensorFactorization.Core.:×₃

BlockTensorFactorization.Core.:×₄

BlockTensorFactorization.Core.:×₅

BlockTensorFactorization.Core.:×₆

BlockTensorFactorization.Core.:×₇

BlockTensorFactorization.Core.:×₈

BlockTensorFactorization.Core.:×₉

BlockTensorFactorization.Core.:⋅₁

BlockTensorFactorization.Core.:⋅₂

BlockTensorFactorization.Core.:⋅₃

BlockTensorFactorization.Core.:⋅₄

BlockTensorFactorization.Core.:⋅₅

BlockTensorFactorization.Core.:⋅₆

BlockTensorFactorization.Core.:⋅₇

BlockTensorFactorization.Core.:⋅₈

BlockTensorFactorization.Core.:⋅₉

BlockTensorFactorization.Core.DEFAULT_INIT

BlockTensorFactorization.Core.Diagonal_col_norm

BlockTensorFactorization.Core.IntervalConstraint

BlockTensorFactorization.Core.ReLU

BlockTensorFactorization.Core._factorize

BlockTensorFactorization.Core._gettuckerindex

BlockTensorFactorization.Core.abs_randn

BlockTensorFactorization.Core.all_recursive

BlockTensorFactorization.Core.array

BlockTensorFactorization.Core.binary!

BlockTensorFactorization.Core.check

BlockTensorFactorization.Core.check

BlockTensorFactorization.Core.circumscribed_radius

BlockTensorFactorization.Core.coarsen

BlockTensorFactorization.Core.contractions

BlockTensorFactorization.Core.core

BlockTensorFactorization.Core.cpproduct

BlockTensorFactorization.Core.cubic_spline_coefficients

BlockTensorFactorization.Core.curvature

BlockTensorFactorization.Core.d2_dx2

BlockTensorFactorization.Core.d_dx

BlockTensorFactorization.Core.d_dx_and_d2_dx2_spline

BlockTensorFactorization.Core.default_kwargs

BlockTensorFactorization.Core.eachfactorindex

BlockTensorFactorization.Core.eachfibre

BlockTensorFactorization.Core.eachrank1term

BlockTensorFactorization.Core.eachrank1term

BlockTensorFactorization.Core.eachrank1term

BlockTensorFactorization.Core.expand_decomposition_constraints

BlockTensorFactorization.Core.factor

BlockTensorFactorization.Core.factorize

BlockTensorFactorization.Core.factors

BlockTensorFactorization.Core.false_tuple

BlockTensorFactorization.Core.finalconstrain!

BlockTensorFactorization.Core.frozen

BlockTensorFactorization.Core.geomean

BlockTensorFactorization.Core.getnotindex

BlockTensorFactorization.Core.group_by_factor

BlockTensorFactorization.Core.identity_tensor

BlockTensorFactorization.Core.identityslice

BlockTensorFactorization.Core.initialize

BlockTensorFactorization.Core.initialize_continuous_dims

BlockTensorFactorization.Core.initialize_decomposition

BlockTensorFactorization.Core.initialize_parameters

BlockTensorFactorization.Core.initialize_previous

BlockTensorFactorization.Core.initialize_scales

BlockTensorFactorization.Core.initialize_stats

BlockTensorFactorization.Core.interlace

BlockTensorFactorization.Core.interpolate

BlockTensorFactorization.Core.isfrozen

BlockTensorFactorization.Core.isnonnegative

BlockTensorFactorization.Core.isnonnegative_sumtoone

BlockTensorFactorization.Core.khatrirao

BlockTensorFactorization.Core.l1normalize!

BlockTensorFactorization.Core.l1normalize_12slices!

BlockTensorFactorization.Core.l1normalize_1slices!

BlockTensorFactorization.Core.l1normalize_cols!

BlockTensorFactorization.Core.l1normalize_rows!

BlockTensorFactorization.Core.l1scale!

BlockTensorFactorization.Core.l1scale_12slices!

BlockTensorFactorization.Core.l1scale_1slices!

BlockTensorFactorization.Core.l1scale_average12slices!

BlockTensorFactorization.Core.l1scale_cols!

BlockTensorFactorization.Core.l1scale_rows!

BlockTensorFactorization.Core.l2scale!

BlockTensorFactorization.Core.l2scale_12slices!

BlockTensorFactorization.Core.l2scale_1slices!

BlockTensorFactorization.Core.l2scale_average12slices!

BlockTensorFactorization.Core.l2scale_cols!

BlockTensorFactorization.Core.l2scale_rows!

BlockTensorFactorization.Core.linftynorm

BlockTensorFactorization.Core.linftynormalize!

BlockTensorFactorization.Core.linftynormalize_12slices!

BlockTensorFactorization.Core.linftynormalize_1slices!

BlockTensorFactorization.Core.linftynormalize_cols!

BlockTensorFactorization.Core.linftynormalize_inftyslices!

BlockTensorFactorization.Core.linftynormalize_rows!

BlockTensorFactorization.Core.linftyproject!

BlockTensorFactorization.Core.linftyscale!

BlockTensorFactorization.Core.linftyscale_12slices!

BlockTensorFactorization.Core.linftyscale_1slices!

BlockTensorFactorization.Core.linftyscale_average12slices!

BlockTensorFactorization.Core.linftyscale_cols!

BlockTensorFactorization.Core.linftyscale_rows!

BlockTensorFactorization.Core.make_spline

BlockTensorFactorization.Core.make_update!

BlockTensorFactorization.Core.mat

BlockTensorFactorization.Core.match_cols!

BlockTensorFactorization.Core.match_factors!

BlockTensorFactorization.Core.match_rows!

BlockTensorFactorization.Core.match_slices!

BlockTensorFactorization.Core.matrix_factor

BlockTensorFactorization.Core.matrix_factors

BlockTensorFactorization.Core.mtt

BlockTensorFactorization.Core.multifoldl

BlockTensorFactorization.Core.multiscale_factorize

BlockTensorFactorization.Core.nfactors

BlockTensorFactorization.Core.nmode_product

BlockTensorFactorization.Core.nmode_product

BlockTensorFactorization.Core.nmp

BlockTensorFactorization.Core.nonnegative!

BlockTensorFactorization.Core.norm2

BlockTensorFactorization.Core.parse_constraints

BlockTensorFactorization.Core.possible_ranks

BlockTensorFactorization.Core.postprocess!

BlockTensorFactorization.Core.proj_one_hot

BlockTensorFactorization.Core.proj_one_hot!

BlockTensorFactorization.Core.projsplx

BlockTensorFactorization.Core.projsplx!

BlockTensorFactorization.Core.rank_detect_factorize

BlockTensorFactorization.Core.rankof

BlockTensorFactorization.Core.rankof

BlockTensorFactorization.Core.reshape_ndims

BlockTensorFactorization.Core.scale_constraint

BlockTensorFactorization.Core.scale_constraints

BlockTensorFactorization.Core.scale_decomposition_constraint

BlockTensorFactorization.Core.simplex!

BlockTensorFactorization.Core.simplex_12slices!

BlockTensorFactorization.Core.simplex_1slices!

BlockTensorFactorization.Core.simplex_cols!

BlockTensorFactorization.Core.simplex_rows!

BlockTensorFactorization.Core.slicewise_dot

BlockTensorFactorization.Core.smart_insert!

BlockTensorFactorization.Core.smart_interlace!

BlockTensorFactorization.Core.spline_mat

BlockTensorFactorization.Core.standard_curvature

BlockTensorFactorization.Core.swapdims

BlockTensorFactorization.Core.tucker_contractions

BlockTensorFactorization.Core.tuckerproduct

BlockTensorFactorization.DensityEstimationTools._in2drange

BlockTensorFactorization.DensityEstimationTools._inrange

BlockTensorFactorization.DensityEstimationTools.coordzip

BlockTensorFactorization.DensityEstimationTools.default_bandwidth

BlockTensorFactorization.DensityEstimationTools.filter_2d_inner_percentile

BlockTensorFactorization.DensityEstimationTools.filter_inner_percentile

BlockTensorFactorization.DensityEstimationTools.kde2d

BlockTensorFactorization.DensityEstimationTools.make_densities

BlockTensorFactorization.DensityEstimationTools.make_densities2d

BlockTensorFactorization.DensityEstimationTools.repeatcoord

BlockTensorFactorization.DensityEstimationTools.standardize_2d_KDEs

BlockTensorFactorization.DensityEstimationTools.standardize_KDEs

BlockTensorFactorization.DensityEstimationTools.standardize_KDEs

BlockTensorFactorization.MatrixTensorFactor._avg_fibre_normalize!

BlockTensorFactorization.MatrixTensorFactor._init

BlockTensorFactorization.MatrixTensorFactor._slice_normalize!

BlockTensorFactorization.MatrixTensorFactor.nnmtf

BlockTensorFactorization.MatrixTensorFactor.rescaleAB!