Matrix Tensor Factor

Old interface from MatrixTensorFactor, with the improved backend from factorize.

Warning

This submodule is deprecated and may be removed in future versions. It is preferred to use factorize rather than nnmtf.

Density Estimation Tools

Matrix Tensor Factor

Block Tensor Decomposition

About this package

Quick Guide

Factorizing your data

Extracting factors

Iteration Statistics

Output keyword arguments

Constants

Functions

Types

Block Update Order

Randomizing Block Updates Order

Constrained Factorization

Example

Ready-to-use Constraints

Advanced Constraints

Manual Constraint Updates

Decomposition Models

How Julia treats an AbstractDecomposition

Iteration Stats

BlockTensorFactorization.MatrixTensorFactor — Module

Matrix-Tensor Factorization

BlockTensorFactorization.MatrixTensorFactor.IMPLIMENTED_CRITERIA — Constant

IMPLIMENTED_CRITERIA::Set{Symbol}

:ncone: vector-set distance between the -gradient of the objective and the normal cone

:iterates: A,B before and after one iteration are close in L2 norm

:objective: objective is small

:relativeerror: relative error is small (when normalize=:nothing) or mean relative error averaging fibres or slices when the normalization is :fibres or :slices respectfuly.

BlockTensorFactorization.MatrixTensorFactor.IMPLIMENTED_METRICS — Constant

IMPLIMENTED_METRICS::Set{Symbol}

:L1: the default; ensure sums of entries in each fibre or slice (according to normalize) are equal to 1

TODO :L2: ensure the sum of squares of entries are 1

:Linfty: ensures the maximum entry is 1

:nothing: do not enforce a metric to equal 1

BlockTensorFactorization.MatrixTensorFactor.IMPLIMENTED_NORMALIZATIONS — Constant

IMPLIMENTED_NORMALIZATIONS::Set{Symbol}

:fibres: set $\sum_{k=1}^K B[r,j,k] = 1$ for all $r, j$, or when projection==:nnscale, set $\sum_{j=1}^J\sum_{k=1}^K B[r,j,k] = J$ for all $r$

:slices: set $\sum_{j=1}^J\sum_{k=1}^K B[r,j,k] = 1$ for all $r$

:nothing: does not enforce any normalization of B

:rows: set a metric on eachslice(X; dims=1) to 1, which is equivelent to eachrow(X) when X is a matrix

:cols: set a metric on eachslice(X; dims=2) to 1, which is equivelent to eachcol(X) when X is a matrix

BlockTensorFactorization.MatrixTensorFactor.IMPLIMENTED_OPTIONS — Constant

Lists all implimented options

BlockTensorFactorization.MatrixTensorFactor.IMPLIMENTED_PROJECTIONS — Constant

IMPLIMENTED_PROJECTIONS::Set{Symbol}

:nnscale: Two stage block coordinate decent; 1) projected gradient decent onto nonnegative orthant, 2) shift any weight from B to A according to normalization. Equivilent to :nonnegative when normalization==:nothing.

:simplex: Euclidian projection onto the simplex blocks accoring to normalization

:nonnegative: zero out negative entries

BlockTensorFactorization.MatrixTensorFactor.IMPLIMENTED_STEPSIZES — Constant

IMPLIMENTED_STEPSIZES::Set{Symbol}

:lipschitz: gradient step 1/L for lipschitz constant L

:spg: spectral projected gradient stepsize

BlockTensorFactorization.MatrixTensorFactor._avg_fibre_normalize! — Method

Rescales A and B so each factor (3 fibres) of B has similar magnitude.

BlockTensorFactorization.MatrixTensorFactor._init — Method

Default initialization

BlockTensorFactorization.MatrixTensorFactor._slice_normalize! — Method

Rescales A and B so each factor (horizontal slices) of B has similar magnitude.

BlockTensorFactorization.MatrixTensorFactor.nnmtf — Function

nnmtf(Y::AbstractArray, R::Integer; kwargs...)

Non-negatively matrix-tensor factorizes an order N tensor Y with a given "rank" R.

For an order $N=3$ tensor, this factorizes $Y \approx A B$ where $\displaystyle Y[i,j,k] \approx \sum_{r=1}^R A[i,r]*B[r,j,k]$ and the factors $A, B \geq 0$ are nonnegative.

For higher orders, this becomes $\displaystyle Y[i1,i2,...,iN] \approx \sum_{r=1}^R A[i1,r]*B[r,i2,...,iN].$

Note there may NOT be a unique optimal solution

Arguments

Y::AbstractArray{T,N}: tensor to factorize

R::Integer: rank to factorize Y (size(A, 2) and size(B, 1))

Keywords

maxiter::Integer=100: maxmimum number of iterations

tol::Real=1e-3: desiered tolerance for the convergence criterion

rescale_AB::Bool=true: scale B at each iteration so that the factors (horizontal slices) have similar 3-fiber sums.

rescale_Y::Bool=true: Preprocesses the input Y to have normalized 3-fiber sums (on average), and rescales the final B so Y=A*B.

normalize::Symbol=:fibres: part of B that should be normalized (must be in IMPLIMENTED_NORMALIZATIONS)

normalizeA::Symbol=:rows: part of A that should be normalized (must be in IMPLIMENTED_NORMALIZATIONS)

projection::Symbol=:nnscale: constraint to use and method for enforcing it (must be in IMPLIMENTED_PROJECTIONS)

metric::Symbol=:L1: under what metric the fibres/slices are normalized (must be in IMPLIMENTED_METRICS)

criterion::Symbol=:ncone: how to determine if the algorithm has converged (must be in IMPLIMENTED_CRITERIA)

stepsize::Symbol=:lipschitz: used for the gradient decent step (must be in IMPLIMENTED_STEPSIZES)

momentum::Bool=false: use momentum updates

delta::Real=0.9999: safeguard for maximum amount of momentum (see eq (3.5) Xu & Yin 2013)

R_max::Integer=size(Y, 1): maximum rank to try if R is not given

projectionA::Symbol=projection: projection to use on factor A (must be in IMPLIMENTED_PROJECTIONS)

projectionB::Symbol=projection: projection to use on factor B (must be in IMPLIMENTED_PROJECTIONS)

metricA::Symbol=metric: the metric to use for factor A (must be in IMPLIMENTED_METRICS)

metricB::Symbol=metric: the metric to use for factor B (must be in IMPLIMENTED_METRICS)

scaleBtoA::Bool=true: when using projection=:nnscale, if the weights should be moved from B to A, or A to B

A_init::AbstractMatrix=nothing: initial A for the iterative algorithm. Should be kept as nothing if R is not given.

B_init::AbstractArray=nothing: initial B for the iterative algorithm. Should be kept as nothing if R is not given.

Returns

A::Matrix{Float64}: the matrix A in the factorization Y ≈ A * B

B::Array{Float64, N}: the tensor B in the factorization Y ≈ A * B

rel_errors::Vector{Float64}: relative errors at each iteration

norm_grad::Vector{Float64}: norm of the full gradient at each iteration

dist_Ncone::Vector{Float64}: distance of the -gradient to the normal cone at each iteration

If R was estimated, also returns the optimal R::Integer

Implimentation of block coordinate decent updates

We calculate the partial gradients and corresponding Lipschitz constants like so:

\[\begin{align} \boldsymbol{P}^{t}[q,r] &=\textstyle{\sum}_{jk} \boldsymbol{\mathscr{B}}^n[q,j,k] \boldsymbol{\mathscr{B}}^n[r,j,k]\\ \boldsymbol{Q}^{t}[i,r] &=\textstyle{\sum}_{jk}\boldsymbol{\mathscr{Y}}[i,j,k] \boldsymbol{\mathscr{B}}^n[r,j,k] \\ \nabla_{A} f(\boldsymbol{A}^{t},\boldsymbol{\mathscr{B}}^{t}) &= \boldsymbol{A}^{t} \boldsymbol{P}^{t} - \boldsymbol{Q}^{t} \\ L_{A} &= \left\lVert \boldsymbol{P}^{t} \right\rVert_{2}. \end{align}\]

Similarly for $\boldsymbol{\mathscr{B}}$:

\[\begin{align} \boldsymbol{T}^{t+1}&=(\boldsymbol{A}^{t+\frac12})^\top \boldsymbol{A}^{t+\frac12}\\ \boldsymbol{\mathscr{U}}^{t+1}&=(\boldsymbol{A}^{t+\frac12})^\top \boldsymbol{\mathscr{Y}} \\ \nabla_\boldsymbol{\mathscr{B}} f(\boldsymbol{A}^{t+\frac12},\boldsymbol{\mathscr{B}}^{t}) &= \boldsymbol{T}^{t+1} \boldsymbol{\mathscr{B}}^{t} - \boldsymbol{\mathscr{U}}^{t+1} \\ L_B &= \left\lVert \boldsymbol{T}^{t+1} \right\rVert_{2}. \end{align}\]

To ensure the iterates stay "close" to normalized, we introduce a renormalization step after the projected gradient updates:

\[\begin{align} \boldsymbol{C} [r,r]&=\frac{1}{J}\textstyle{\sum}_{jk} \boldsymbol{\mathscr{B}}^{t+\frac12}[r,j,k]\\ \boldsymbol{A}^{t+1}&= \boldsymbol{A}^{t+\frac12} \boldsymbol{C}\\ \boldsymbol{\mathscr{B}}^{t+1}&= (\boldsymbol{C}^{t+1})^{-1}\boldsymbol{\mathscr{B}}^{t+\frac12}. \end{align}\]

We typicaly use the following convergence criterion:

\[d(-\nabla \ell(\boldsymbol{A}^{t},\boldsymbol{\mathscr{B}}^{t}), N_{\mathcal{C}}(\boldsymbol{A}^{t},\boldsymbol{\mathscr{B}}^{t}))^2\leq\delta^2 R(I+JK).\]

BlockTensorFactorization.MatrixTensorFactor.rescaleAB! — Method

Rescales A and B so each factor (horizontal slices) of B has similar magnitude.

BlockTensorFactorization.MatrixTensorFactor

BlockTensorFactorization.DensityEstimationTools.DEFAULT_ALPHA

BlockTensorFactorization.DensityEstimationTools.DEFAULT_N_SAMPLES

BlockTensorFactorization.MatrixTensorFactor.IMPLIMENTED_CRITERIA

BlockTensorFactorization.MatrixTensorFactor.IMPLIMENTED_METRICS

BlockTensorFactorization.MatrixTensorFactor.IMPLIMENTED_NORMALIZATIONS

BlockTensorFactorization.MatrixTensorFactor.IMPLIMENTED_OPTIONS

BlockTensorFactorization.MatrixTensorFactor.IMPLIMENTED_PROJECTIONS

BlockTensorFactorization.MatrixTensorFactor.IMPLIMENTED_STEPSIZES

BlockTensorFactorization.Core.AbstractConstraint

BlockTensorFactorization.Core.AbstractDecomposition

BlockTensorFactorization.Core.AbstractObjective

BlockTensorFactorization.Core.AbstractStat

BlockTensorFactorization.Core.AbstractStep

BlockTensorFactorization.Core.AbstractTucker

BlockTensorFactorization.Core.BlockGradientDescent

BlockTensorFactorization.Core.CPDecomposition

BlockTensorFactorization.Core.ComposedConstraint

BlockTensorFactorization.Core.ConstraintUpdate

BlockTensorFactorization.Core.DiagonalTuple

BlockTensorFactorization.Core.DisplayDecomposition

BlockTensorFactorization.Core.Entrywise

BlockTensorFactorization.Core.Entrywise

BlockTensorFactorization.Core.EuclidianLipschitz

BlockTensorFactorization.Core.EuclidianStepSize

BlockTensorFactorization.Core.FactorNorms

BlockTensorFactorization.Core.GenericConstraint

BlockTensorFactorization.Core.GenericDecomposition

BlockTensorFactorization.Core.GradientDescent

BlockTensorFactorization.Core.L2

BlockTensorFactorization.Core.L2

BlockTensorFactorization.Core.LinearConstraint

BlockTensorFactorization.Core.LipschitzStep

BlockTensorFactorization.Core.LipschitzStep

BlockTensorFactorization.Core.MomentumUpdate

BlockTensorFactorization.Core.NoConstraint

BlockTensorFactorization.Core.PrintStats

BlockTensorFactorization.Core.ProjectedNormalization

BlockTensorFactorization.Core.Projection

BlockTensorFactorization.Core.Rescale

BlockTensorFactorization.Core.ScaledNormalization

BlockTensorFactorization.Core.SingletonDecomposition

BlockTensorFactorization.Core.SuperDiagonal

BlockTensorFactorization.Core.SuperDiagonal

BlockTensorFactorization.Core.Tucker

BlockTensorFactorization.Core.Tucker

BlockTensorFactorization.Core.Tucker1

BlockTensorFactorization.Core.Tucker1

BlockTensorFactorization.Core.:×₁

BlockTensorFactorization.Core.:×₂

BlockTensorFactorization.Core.:×₃

BlockTensorFactorization.Core.:×₄

BlockTensorFactorization.Core.:×₅

BlockTensorFactorization.Core.:×₆

BlockTensorFactorization.Core.:×₇

BlockTensorFactorization.Core.:×₈

BlockTensorFactorization.Core.:×₉

BlockTensorFactorization.Core.:⋅₁

BlockTensorFactorization.Core.:⋅₂

BlockTensorFactorization.Core.:⋅₃

BlockTensorFactorization.Core.:⋅₄

BlockTensorFactorization.Core.:⋅₅

BlockTensorFactorization.Core.:⋅₆

BlockTensorFactorization.Core.:⋅₇

BlockTensorFactorization.Core.:⋅₈

BlockTensorFactorization.Core.:⋅₉

BlockTensorFactorization.Core.DEFAULT_INIT

BlockTensorFactorization.Core.Diagonal_col_norm

BlockTensorFactorization.Core.IntervalConstraint

BlockTensorFactorization.Core.ReLU

BlockTensorFactorization.Core._factorize

BlockTensorFactorization.Core._gettuckerindex

BlockTensorFactorization.Core.abs_randn

BlockTensorFactorization.Core.all_recursive

BlockTensorFactorization.Core.array

BlockTensorFactorization.Core.binary!

BlockTensorFactorization.Core.check

BlockTensorFactorization.Core.check

BlockTensorFactorization.Core.circumscribed_radius

BlockTensorFactorization.Core.coarsen

BlockTensorFactorization.Core.contractions

BlockTensorFactorization.Core.core

BlockTensorFactorization.Core.cpproduct

BlockTensorFactorization.Core.cubic_spline_coefficients

BlockTensorFactorization.Core.curvature

BlockTensorFactorization.Core.d2_dx2

BlockTensorFactorization.Core.d_dx

BlockTensorFactorization.Core.d_dx_and_d2_dx2_spline

BlockTensorFactorization.Core.default_kwargs

BlockTensorFactorization.Core.eachfactorindex

BlockTensorFactorization.Core.eachfibre

BlockTensorFactorization.Core.eachrank1term

BlockTensorFactorization.Core.eachrank1term

BlockTensorFactorization.Core.eachrank1term

BlockTensorFactorization.Core.expand_decomposition_constraints

BlockTensorFactorization.Core.factor

BlockTensorFactorization.Core.factorize

BlockTensorFactorization.Core.factors

BlockTensorFactorization.Core.false_tuple

BlockTensorFactorization.Core.finalconstrain!

BlockTensorFactorization.Core.frozen

BlockTensorFactorization.Core.geomean

BlockTensorFactorization.Core.getnotindex

BlockTensorFactorization.Core.group_by_factor

BlockTensorFactorization.Core.identity_tensor

BlockTensorFactorization.Core.identityslice

BlockTensorFactorization.Core.initialize

BlockTensorFactorization.Core.initialize_continuous_dims

BlockTensorFactorization.Core.initialize_decomposition

BlockTensorFactorization.Core.initialize_parameters

BlockTensorFactorization.Core.initialize_previous

BlockTensorFactorization.Core.initialize_scales

BlockTensorFactorization.Core.initialize_stats

BlockTensorFactorization.Core.interlace

BlockTensorFactorization.Core.interpolate

BlockTensorFactorization.Core.isfrozen

BlockTensorFactorization.Core.isnonnegative

BlockTensorFactorization.Core.isnonnegative_sumtoone

BlockTensorFactorization.Core.khatrirao

BlockTensorFactorization.Core.l1normalize!

BlockTensorFactorization.Core.l1normalize_12slices!

BlockTensorFactorization.Core.l1normalize_1slices!

BlockTensorFactorization.Core.l1normalize_cols!

BlockTensorFactorization.Core.l1normalize_rows!

BlockTensorFactorization.Core.l1scale!

BlockTensorFactorization.Core.l1scale_12slices!

BlockTensorFactorization.Core.l1scale_1slices!

BlockTensorFactorization.Core.l1scale_average12slices!

BlockTensorFactorization.Core.l1scale_cols!

BlockTensorFactorization.Core.l1scale_rows!

BlockTensorFactorization.Core.l2scale!

BlockTensorFactorization.Core.l2scale_12slices!

BlockTensorFactorization.Core.l2scale_1slices!

BlockTensorFactorization.Core.l2scale_average12slices!

BlockTensorFactorization.Core.l2scale_cols!

BlockTensorFactorization.Core.l2scale_rows!

BlockTensorFactorization.Core.linftynorm

BlockTensorFactorization.Core.linftynormalize!

BlockTensorFactorization.Core.linftynormalize_12slices!

BlockTensorFactorization.Core.linftynormalize_1slices!

BlockTensorFactorization.Core.linftynormalize_cols!

BlockTensorFactorization.Core.linftynormalize_inftyslices!

BlockTensorFactorization.Core.linftynormalize_rows!

BlockTensorFactorization.Core.linftyproject!

BlockTensorFactorization.Core.linftyscale!

BlockTensorFactorization.Core.linftyscale_12slices!

BlockTensorFactorization.Core.linftyscale_1slices!

BlockTensorFactorization.Core.linftyscale_average12slices!

BlockTensorFactorization.Core.linftyscale_cols!

BlockTensorFactorization.Core.linftyscale_rows!

BlockTensorFactorization.Core.make_spline

BlockTensorFactorization.Core.make_update!

BlockTensorFactorization.Core.mat

BlockTensorFactorization.Core.match_cols!

BlockTensorFactorization.Core.match_factors!

BlockTensorFactorization.Core.match_rows!

BlockTensorFactorization.Core.match_slices!

BlockTensorFactorization.Core.matrix_factor

BlockTensorFactorization.Core.matrix_factors

BlockTensorFactorization.Core.mtt

BlockTensorFactorization.Core.multifoldl

BlockTensorFactorization.Core.multiscale_factorize

BlockTensorFactorization.Core.nfactors

BlockTensorFactorization.Core.nmode_product

BlockTensorFactorization.Core.nmode_product

BlockTensorFactorization.Core.nmp

BlockTensorFactorization.Core.nonnegative!

BlockTensorFactorization.Core.norm2

BlockTensorFactorization.Core.parse_constraints

BlockTensorFactorization.Core.possible_ranks

BlockTensorFactorization.Core.postprocess!

BlockTensorFactorization.Core.proj_one_hot

BlockTensorFactorization.Core.proj_one_hot!

BlockTensorFactorization.Core.projsplx

BlockTensorFactorization.Core.projsplx!

BlockTensorFactorization.Core.rank_detect_factorize

BlockTensorFactorization.Core.rankof

BlockTensorFactorization.Core.rankof

BlockTensorFactorization.Core.reshape_ndims

BlockTensorFactorization.Core.scale_constraint

BlockTensorFactorization.Core.scale_constraints

BlockTensorFactorization.Core.scale_decomposition_constraint

BlockTensorFactorization.Core.simplex!

BlockTensorFactorization.Core.simplex_12slices!

BlockTensorFactorization.Core.simplex_1slices!

BlockTensorFactorization.Core.simplex_cols!

BlockTensorFactorization.Core.simplex_rows!

BlockTensorFactorization.Core.slicewise_dot

BlockTensorFactorization.Core.smart_insert!

BlockTensorFactorization.Core.smart_interlace!

BlockTensorFactorization.Core.spline_mat

BlockTensorFactorization.Core.standard_curvature

BlockTensorFactorization.Core.swapdims

BlockTensorFactorization.Core.tucker_contractions

BlockTensorFactorization.Core.tuckerproduct

BlockTensorFactorization.DensityEstimationTools._in2drange

BlockTensorFactorization.DensityEstimationTools._inrange

BlockTensorFactorization.DensityEstimationTools.coordzip

BlockTensorFactorization.DensityEstimationTools.default_bandwidth

BlockTensorFactorization.DensityEstimationTools.filter_2d_inner_percentile

BlockTensorFactorization.DensityEstimationTools.filter_inner_percentile

BlockTensorFactorization.DensityEstimationTools.kde2d

BlockTensorFactorization.DensityEstimationTools.make_densities

BlockTensorFactorization.DensityEstimationTools.make_densities2d

BlockTensorFactorization.DensityEstimationTools.repeatcoord

BlockTensorFactorization.DensityEstimationTools.standardize_2d_KDEs

BlockTensorFactorization.DensityEstimationTools.standardize_KDEs

BlockTensorFactorization.DensityEstimationTools.standardize_KDEs

BlockTensorFactorization.MatrixTensorFactor._avg_fibre_normalize!

BlockTensorFactorization.MatrixTensorFactor._init

BlockTensorFactorization.MatrixTensorFactor._slice_normalize!

BlockTensorFactorization.MatrixTensorFactor.nnmtf

BlockTensorFactorization.MatrixTensorFactor.rescaleAB!