Iteration Stats

BlockTensorFactorization.Core.AbstractStat — Type

An AbstractStat is a type which, when created, can be applied to the four arguments (X::AbstractDecomposition, Y::AbstractArray, previous::Vector{<:AbstractDecomposition}, parameters::Dict) to (usually) return a number.

The following stats are supported inputs to the stats keyword in factorize.

Iteration
GradientNorm
GradientNNCone
ObjectiveValue
ObjectiveRatio
RelativeError
IterateNormDiff
IterateRelativeDiff
EuclidianStepSize
EuclidianLipschitz
FactorNorms
PrintStats
DisplayDecomposition

Detail

BlockTensorFactorization.Core.Iteration — Type

Iteration <: AbstractStat

Iteration number.

BlockTensorFactorization.Core.GradientNorm — Type

GradientNorm{T} <: AbstractStat

2-norm of the gradient.

BlockTensorFactorization.Core.GradientNNCone — Type

GradientNNCone{T} <: AbstractStat

2-norm vector-set distance between the negative gradient and nonnegative cone at the iterate.

BlockTensorFactorization.Core.ObjectiveValue — Type

ObjectiveValue{T<:AbstractObjective} <: AbstractStat

The current objective value.

BlockTensorFactorization.Core.ObjectiveRatio — Type

ObjectiveRatio{T<:AbstractObjective} <: AbstractStat

Ratio between the previous and current objective value.

BlockTensorFactorization.Core.RelativeError — Type

RelativeError{T<:Function} <: AbstractStat

Relative error between the decomposition model, and input array.

BlockTensorFactorization.Core.IterateNormDiff — Type

IterateNormDiff{T<:Function} <: AbstractStat

2-norm of the difference between the previous and current iterate.

BlockTensorFactorization.Core.IterateRelativeDiff — Type

IterateRelativeDiff{T<:Function} <: AbstractStat

Relative difference between the previous and current iterate.

BlockTensorFactorization.Core.EuclidianStepSize — Type

The 2-norm of the stepsizes that would be taken for all blocks.

For example, if there are two blocks, and we would take a stepsize of A to update one block and B to update the other, this would return sqrt(A^2 + B^2).

BlockTensorFactorization.Core.EuclidianLipschitz — Type

The 2-norm of the lipschitz constants that would be taken for all blocks.

Need the stepsizes to be lipschitz steps since it is calculated similarly to EuclidianStepSize.

BlockTensorFactorization.Core.FactorNorms — Type

FactorNorms(; norm, kwargs...)

Makes a tuple containing the norm of each factor in the decomposition.

Auxiliary Stats

The following are subtype of AbstractStat but are for auxiliary features.

BlockTensorFactorization.Core.PrintStats — Type

PrintStats(; kwargs...)

Does not use any of the kwargs. Simply prints the most recent row of the stats.

BlockTensorFactorization.Core.DisplayDecomposition — Type

DisplayDecomposition(; kwargs...)

Does not use any of the kwargs. Simply displays the current iteration.