Density Estimation Tools

Helper functions for working with data for density tensors.

Warning

This submodule is deprecated and may be removed in future versions.

BlockTensorFactorization.DensityEstimationTools.DEFAULT_ALPHA — Constant

DEFAULT_ALPHA = 0.9::Real

Smoothing parameter for calculating a kernel's bandwidth.

BlockTensorFactorization.DensityEstimationTools.DEFAULT_N_SAMPLES — Constant

DEFAULT_N_SAMPLES = 64::Int

Number of samples to use when standardizing a vector of density estimates.

BlockTensorFactorization.DensityEstimationTools._in2drange — Method

Returns a function that checks if each coordinate is in the inner P percentile of the values in vs.

BlockTensorFactorization.DensityEstimationTools._inrange — Method

Returns a function that checks if a value is in the inner P percentile of the values in v.

BlockTensorFactorization.DensityEstimationTools.coordzip — Method

coordzip(rcoords)

Zips the "x" and "y" values together into a list of x coords and y coords. Example –––- coordzip([(0,0), (1,1), (1,1), (1,1), (1,2), (1,2)])

[[0, 1, 1, 1, 1, 1], [0, 1, 1, 1, 2, 2]]

BlockTensorFactorization.DensityEstimationTools.default_bandwidth — Function

default_bandwidth(data; alpha=0.9, inner_percentile=100)

Coppied from KernelDensity since this function is not exported. I want access to it so that the same bandwidth can be used for different densities for the same measurements.

BlockTensorFactorization.DensityEstimationTools.filter_2d_inner_percentile — Method

Filters 2d elements so only the ones in the inner P percentile remain. See filter_inner_percentile.

BlockTensorFactorization.DensityEstimationTools.filter_inner_percentile — Method

Filters elements so only the ones in the inner P percentile remain. See filter_2d_inner_percentile.

BlockTensorFactorization.DensityEstimationTools.kde2d — Method

kde2d((xs, ys), values)

Performs a 2d KDE based on two lists of coordinates, and the value at those coordinates. Input ––-

xs, ys::Vector{Real}: coordinates/locations of samples

values::Vector{Integer}: value of the sample

Returns

f::BivariateKDE use f.x, f.y for the location of the (re)sampled KDE,

and f.density for the sample values of the KDE

BlockTensorFactorization.DensityEstimationTools.make_densities — Method

make_densities(s::Sink; kwargs...)
make_densities(s::Sink, domains::AbstractVector{<:AbstractVector}; kwargs...)

Estimates the densities for each measurement in a Sink.

When given domains, a list where each entry is a domain for a different measurement, resample the kernel on this domain.

Parameters

bandwidths::AbstractVector{<:Real}: list of bandwidths used for each measurement's

density estimation

inner_percentile::Integer=100: value between 0 and 100 that filters out each measurement

by using the inner percentile range. This can help remove outliers and focus in on where the bulk of the data is.

Returns

density_estimates::Vector{UnivariateKDE}

BlockTensorFactorization.DensityEstimationTools.make_densities2d — Method

makedensities2d(s::Sink; kwargs...) makedensities2d(s::Sink, domains::AbstractVector{<:AbstractVector}; kwargs...)

Similar to make_densities but performs the KDE on 2 measurements jointly.

BlockTensorFactorization.DensityEstimationTools.repeatcoord — Method

repeatcoord(coordinates, values)

Repeates coordinates the number of times given by values.

Both lists should be the same length.

Example

coordinates = [(0,0), (1,1), (1,2)] values = [1, 3, 2] repeatcoord(coordinates, values)

[(0,0), (1,1), (1,1), (1,1), (1,2), (1,2)]

BlockTensorFactorization.DensityEstimationTools.standardize_2d_KDEs — Method

standardize_2d_KDEs(KDEs::AbstractVector{BivariateKDE}; n_samples=DEFAULT_N_SAMPLES,)

Resample the densities so they all are sampled from the same x and y coordinates.

BlockTensorFactorization.DensityEstimationTools.standardize_KDEs — Method

Resample the densities within each sink/source so that like-measurements use the same scale.

BlockTensorFactorization.DensityEstimationTools.standardize_KDEs — Method

standardize_KDEs(KDEs::AbstractVector{UnivariateKDE}; n_samples=DEFAULT_N_SAMPLES,)

Resample the densities so they all are sampled from the same domain.